+++ lib/tasks/label_update_task.dart
@@ -1,40 +1,45 @@
 import 'dart:io';
+import '../rebrand_config.dart';
 import 'rebrand_task.dart';
 
 class LabelUpdateTask extends RebrandTask {
   @override
-  String get name => "Updating App Name to '$appName'";
+  String get name => "Updating App Name to '${config.appName}'";
 
-  final String appName;
+  final RebrandConfig config;
 
-  LabelUpdateTask(this.appName);
+  LabelUpdateTask(this.config);
 
   @override
   Future<void> execute() async {
-    print("🤖 Updating Android App Label...");
-    final manifest = File('android/app/src/main/AndroidManifest.xml');
-    if (manifest.existsSync()) {
-      var content = manifest.readAsStringSync();
-      content = content.replaceAll(
-        RegExp(r'android:label="[^"]*"'),
-        'android:label="$appName"',
-      );
-      manifest.writeAsStringSync(content);
+    if (config.enableAndroid) {
+      print("🤖 Updating Android App Label...");
+      final manifest = File('android/app/src/main/AndroidManifest.xml');
+      if (manifest.existsSync()) {
+        var content = manifest.readAsStringSync();
+        content = content.replaceAll(
+          RegExp(r'android:label="[^"]*"'),
+          'android:label="${config.appName}"',
+        );
+        manifest.writeAsStringSync(content);
+      }
     }
 
-    print("🍎 Updating iOS App Name...");
-    final plist = File('ios/Runner/Info.plist');
-    if (plist.existsSync()) {
-      var content = plist.readAsStringSync();
-      content = content.replaceAll(
-        RegExp(r'<key>CFBundleName</key>\s*<string>[^<]*</string>'),
-        '<key>CFBundleName</key>\n\t<string>$appName</string>',
-      );
-      content = content.replaceAll(
-        RegExp(r'<key>CFBundleDisplayName</key>\s*<string>[^<]*</string>'),
-        '<key>CFBundleDisplayName</key>\n\t<string>$appName</string>',
-      );
-      plist.writeAsStringSync(content);
+    if (config.enableIOS) {
+      print("🍎 Updating iOS App Name...");
+      final plist = File('ios/Runner/Info.plist');
+      if (plist.existsSync()) {
+        var content = plist.readAsStringSync();
+        content = content.replaceAll(
+          RegExp(r'<key>CFBundleName</key>\s*<string>[^<]*</string>'),
+          '<key>CFBundleName</key>\n\t<string>${config.appName}</string>',
+        );
+        content = content.replaceAll(
+          RegExp(r'<key>CFBundleDisplayName</key>\s*<string>[^<]*</string>'),
+          '<key>CFBundleDisplayName</key>\n\t<string>${config.appName}</string>',
+        );
+        plist.writeAsStringSync(content);
+      }
     }
   }
 }